revert MRuby value implementation “TypeObjectUnion"

+++ src/MRubyCS/RInteger.cs
@@ -2,9 +2,9 @@ namespace MRubyCS;
 
 public sealed class RInteger : RObject
 {
-    public nint Value { get; }
+    public long Value { get; }
 
-    internal RInteger(nint value, RClass integerClass) : base(MRubyVType.Integer, integerClass)
+    internal RInteger(long value, RClass integerClass) : base(MRubyVType.Integer, integerClass)
     {
         Value = value;
     }
